Convert window box-sizing to border-box
There's a bunch of code here attempting to accurately compute box sizing under the CSS 2 standard "content-box" rules. That requires meticulously subtracting padding, keeping track of the scrollbar, etc. etc.

There was a comment in the code explaining the problem:

/* The trick is that left/right/top/bottom are measured to the outside
     of the border, but width/height are measured from the inside of the
     border. (Measured by the browser's DOM methods, I mean.) */

The "browser's DOM methods" are the methods specified by the CSS 2 specification, aka the "content-box" model, but IE's "border-box" model was actually better in many ways. Under content-box sizing, the "width" of an element is the width of its content, excluding padding and border. Under border-box sizing, "width" includes the padding and border.

In CSS3, they introduced the "box-sizing" CSS parameter, to opt-in to border-box sizing.

https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing

It's extremely widely supported, even in IE11. https://caniuse.com/css3-boxsizing

(I'm adopting it here and now not just to clean up the code, but because it will make it easier for us to nudge the "top" and "left" of floating grid windows if we don't have to recalculate "bottom" and "right" as well; now accept_one_window can set the "height" and "width", and we can mess around with "top" and "left" whenever we want.)

I've tested this on Windows Chrome, Edge, and Firefox, and on macOS Safari, Chrome, and Firefox.